Ultimos retoques esteticos y funcionales

+++ Controllers/UserController.cs
@@ -31,8 +31,18 @@ namespace CrudApp.Controllers
         [HttpPost]
         public IActionResult Create(User user)
         {
+            if (!ModelState.IsValid)
+            {
+                return View(user);
+            }
+            if (_context.Users.Any(u => u.Username.ToLower() == user.Username.ToLower()))
+            {
+                ModelState.AddModelError("Username", "Este usuario ya existe");
+                return View(user);
+            }
             _context.Users.Add(user);
             _context.SaveChanges();
+
             return RedirectToAction("Index");
         }
 
@@ -47,8 +57,14 @@ namespace CrudApp.Controllers
         [HttpPost]
         public IActionResult Edit(User user)
         {
+            if (!ModelState.IsValid)
+            {
+                return View(user);
+            }
+
             _context.Users.Update(user);
             _context.SaveChanges();
+
             return RedirectToAction("Index");
         }
 
@@ -56,8 +72,28 @@ namespace CrudApp.Controllers
         public IActionResult Delete(int id)
         {
             var user = _context.Users.Find(id);
+
+            if (user == null)
+            {
+                return NotFound(); // evita errores
+            }
+
+            return View(user);
+        }
+
+        [HttpPost, ActionName("Delete")]
+        public IActionResult DeleteConfirmed(int id)
+        {
+            var user = _context.Users.Find(id);
+
+            if (user == null)
+            {
+                return NotFound();
+            }
+
             _context.Users.Remove(user);
             _context.SaveChanges();
+
             return RedirectToAction("Index");
         }
     }
